fix: golint error that delete unused var and unnecessary conversion

+++ util/log/internal/zap/json_encoder.go
@@ -198,7 +198,7 @@ func (enc *zapJsonEncoder) AppendByteString(val []byte) {
 func (enc *zapJsonEncoder) AppendComplex128(val complex128) {
 	enc.addElementSeparator()
 	// Cast to a platform-independent, fixed-size type.
-	r, i := float64(real(val)), float64(imag(val))
+	r, i := real(val), imag(val)
 	enc.buf.AppendByte('"')
 	// Because we're always in a quoted string, we can use strconv without
 	// special-casing NaN and +/-Inf.
